+++ R/add_control_bias.R
@@ -6,15 +6,15 @@
 #' way is to control directly for the bias in the model. Currently supported
 #' methods are:
 #'
-#' [*] \code{"partial"} - An approach described by Warton et al. (2013) to
+#' * \code{"partial"} - An approach described by Warton et al. (2013) to
 #' control the biases in a model, by including a specified variable ("layer") in
 #' the model, but "partialling" it out during the projection phase. Specifically
 #' the variable is set to a specified value ("bias_value"), which is by default
 #' the minimum value observed across the background.
-#' [*] \code{"offset"} - Dummy
+#' * \code{"offset"} - Dummy
 #' method that points to the [`add_offset_bias()`] functionality (see note).
 #' Makes use of offsets to factor out a specified bias variable.
-#' [*] \code{"proximity"} - Use the proximity or distance between points as a weight
+#' * \code{"proximity"} - Use the proximity or distance between points as a weight
 #' in the model. This option effectively places greater weight on points farther
 #' away. *Note:* In the best case this can control for spatial bias and
 #' aggregation, in the worst case it can place a lot of emphasis on points that