+++ NEWS.md
@@ -2,6 +2,7 @@
 
 #### New features
 * Switched object structure to `R6` throughout for improved data and memory handling #44
+* Implemented a convenience function ro remove biodiversity datasets (`rm_biodiversity()`).
 
 #### Minor improvements and bug fixes
 * Added a logical parameter to `ensemble()` enabling compositing of thresholds if set #84